+++ SDGDiscordBot.js
@@ -21,24 +21,31 @@ const AuthDetails = require('./auth.json');
 const MemeManager = require('./modules/memeRepliesModule.js');
 
 const MemeReplies = new MemeManager(bot);
+
 const TableCatchManager = require('./modules/tableCatcherModule.js');
 
 const CatchManager = new TableCatchManager(bot);
+
 const InfoManager = require('./modules/informationModule.js');
 
 const InfoReplies = new InfoManager(bot);
+
 const ServerLogManager = require('./modules/serverLogModule.js');
 
 const ServerLog = new ServerLogManager(bot);
+
 const FinanceManager = require('./modules/financeModule.js');
 
 const FinanceInfo = new FinanceManager(bot);
+
 const FFXIVManager = require('./modules/ffxivModule.js');
 
 const FFXIVInfo = new FFXIVManager(bot);
+
 const SteamManager = require('./modules/steamModule.js');
 
 const steamManager = new SteamManager(bot);
+
 const DrinkingManager = require('./modules/drinkingModule.js');
 
 const drinkingManager = new DrinkingManager(bot);
@@ -53,6 +60,10 @@ MongoClient.connect('mongodb://localhost:27017shamebotdb', { useUnifiedTopology:
   ffxivCollection = db.collection('FFXIV');
   drinkingCollection = db.collection('Drinking');
 });
+
+const HonorManager = require('./modules/honorModule.js');
+
+const honorManager = new HonorManager();
 
 // initialize file transport for winston
 winstonModule.createWinstonFileTransport();
@@ -86,7 +97,6 @@ bot.on('message', (message) => {
   if (message.channel.type === 'dm' || message.channel.type === 'group') {
     return;
   }
-  const serverID = message.channel.guild.id;
 
   // don't listen for self messages
   if (message.author.id === bot.user.id || message.author.bot) {
@@ -96,73 +106,13 @@ bot.on('message', (message) => {
   const lowerCaseMessage = message.content.toLowerCase();
 
   // ------------------------- Honor System Start ------------------------||
-  function returnHonor(user, serverID) {
-    const myCursor = honorCollection.find(
-      { id: user.id, server: serverID },
-      { upvotes: 1, downvotes: 1 },
-    );
-    myCursor.each((err, doc) => {
-      if (err) throw err;
-      if (doc != null) {
-        const { upvotes } = doc;
-        const { downvotes } = doc;
-        const netHonor = upvotes - downvotes;
-        message.reply(`${user.username} has ${netHonor} honor!`);
-      }
-    });
-  }
-
-  function initializeAndReturnHonor(user, serverID, callback) {
-    honorCollection.findOne({ id: user.id, server: serverID }, (err, doc) => {
-      if (err) throw err;
-      if (doc == null) {
-        honorCollection.insertOne(
-          {
-            id: user.id, server: serverID, upvotes: 0, downvotes: 0,
-          },
-        );
-      }
-      callback(user, serverID);
-    });
-  }
-
-  if (message.mentions.users) { // check if message has any mentions
-    // store array of user objects which were mentioned in message
-    const mentionsArray = message.mentions.users.array();
-    const messageTokens = message.content.split(' '); // tokenize message into array
-    for (let i = 0; i < messageTokens.length; i += 1) { // step through tokenized message array
-      if (messageTokens[i].charAt(0) === '<') { // check if word is a mention by checking for the opening char
-        let legitMention = false; // initialize flag for legitimate mentions
-        const slicedStringMention = messageTokens[i].slice(2, -1).toString().replace(/[!]/g, ''); // extract user ID from mention in the message (removing <@ and >) && removes ! from userIDs with a nickname
-        for (const userObj of mentionsArray) { // step through each object in the mentions array
-          // check to see if the user ID found matches a real mention
-          if (userObj.id === slicedStringMention && !userObj.equals(message.author)) {
-            legitMention = true;
-          }
-        }
-        if (legitMention === true) { // if the user id from the string is good
-          if (messageTokens[i + 1] === '++') { // check to see if the token following the mention is a '++' for upvote
-            honorCollection.updateOne(
-              { id: slicedStringMention, server: serverID },
-              { $inc: { upvotes: 1 } },
-              { upsert: true },
-            );
-          } else if (messageTokens[i + 1] === '--' || messageTokens[i + 1] === '—') { // check to see if the token following the mention is a '--' for downvote
-            honorCollection.updateOne(
-              { id: slicedStringMention, server: serverID },
-              { $inc: { downvotes: 1 } },
-              { upsert: true },
-            );
-          }
-        }
-      }
-    }
-  }
-
   // Command to return user's honor
   if (message.content.includes('!honor') && message.mentions.users.array().length === 1) {
-    const mentionedUser = message.mentions.users.array()[0];
-    initializeAndReturnHonor(mentionedUser, serverID, returnHonor);
+    honorManager.initializeAndReturnHonor(message, honorCollection);
+  }
+
+  if (message.mentions.users) {
+    honorManager.updateHonor(message, honorCollection);
   }
 
   // -------------------------- Honor System End -------------------------||
